Resolve CORS issues in Docker, make docker container wait for DB
When ran locally there was no issue but when ran in Docker we were getting CORS issues, probably something to do with what IP the process in the container binds to. It works now anyways. I've essentially nuked cors by telling it to accept any possible URL. Not optimal but it works for testing!